+++ Gabryel_Alves_De_Freitas/SegundoBimestre/pix-api/src/main/java/com/fag/infra/celcoin/dto/CelcoinPixDTO.java
@@ -1,5 +1,7 @@
 package com.fag.infra.celcoin.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.Setter;
 
@@ -7,7 +9,6 @@ import lombok.Setter;
 @Setter
 public class CelcoinPixDTO {
   private String key;
-  private Double amount;
+  private BigDecimal amount;
   private CelcoinMerchantDTO merchant;
-  //private String transactionIdentification;
 }
