+++ Gabryel_Alves_De_Freitas/SegundoBimestre/pix-api/src/main/java/com/fag/infra/panache/model/PanachePixCode.java
@@ -1,5 +1,6 @@
 package com.fag.infra.panache.model;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
@@ -25,7 +26,7 @@ public class PanachePixCode extends PanacheEntityBase {
     private String key;
 
     @Column(name = "AMOUNT")
-    private Double amount;
+    private BigDecimal amount;
 
     public UUID getId() {
         return id;
@@ -59,11 +60,11 @@ public class PanachePixCode extends PanacheEntityBase {
         this.key = key;
     }
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
